+++ src/components/pages/Dashboard.jsx
@@ -1,45 +1,27 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import Header from '../Header';
+import AddEducation from '../AddEducation';
+import AddExperience from '../AddExperience';
+import EditProfile from '../EditProfile';
+import '../../scss/index.scss';
 
 function Dashboard() {
   return (
     <>
       <Header />
       <section className="container">
-        <h1 className="large text-primary">Dashboard</h1>
+        <h1 className="large t-primary">Dashboard</h1>
         <p className="lead">
           <i className="fas fa-user" />
           &nbsp;
           Welcome Jason Doe
         </p>
-        <div className="dash-buttons">
-          <Link
-            to="/dashboard/edit-profile"
-            className="btn"
-          >
-            <i className="fas fa-user-circle text-primary" />
-            &nbsp;
-            Edit Profile
-          </Link>
-          <Link
-            to="/dashboard/add-experience"
-            className="btn"
-          >
-            <i className="fab fa-black-tie text-primary" />
-            &nbsp;
-            Add Experience
-          </Link>
-          <Link
-            to="/dashboard/add-education"
-            className="btn"
-          >
-            <i className="fas fa-graduation-cap text-primary" />
-            &nbsp;
-            Add Education
-          </Link>
+        <div className="dash-buttons mt-4 mb-5">
+          <EditProfile />
+          <AddExperience />
+          <AddEducation />
         </div>
-        <h2 className="my-2">Experience Credentials</h2>
+        <h2 className="mb-4">Experience Credentials</h2>
         <table className="table">
           <thead>
             <tr>
@@ -68,7 +50,7 @@ function Dashboard() {
             </tr>
           </tbody>
         </table>
-        <h2 className="my-2">Education Credentials</h2>
+        <h2 className="mt-4 mb-4">Education Credentials</h2>
         <table className="table">
           <thead>
             <tr>
@@ -89,7 +71,7 @@ function Dashboard() {
             </tr>
           </tbody>
         </table>
-        <div className="my-2">
+        <div className="my-5">
           <button
             type="button"
             className="btn btn-danger"
