+++ src/components/pages/Developers.jsx
@@ -9,7 +9,7 @@ function Developers() {
     <>
       <Header />
       <section className="container">
-        <h1 className="large text-primary">Developers</h1>
+        <h1 className="large t-primary">Developers</h1>
         <p className="lead">
           <i className="fab fa-connectdevelop" />
           &nbsp;
